requirements.txt
asgiref==3.4.1
Django==3.2.6
django-appconf==1.0.5
django-compressor==4.1
django-crispy-forms==1.14.0
django-libsass==0.9
django-markdownx==3.0.1
django-taggit==1.5.1
libsass==0.22.0
Markdown==3.4.1
Pillow==9.3.0
pytz==2021.1
rcssmin==1.1.0
rjsmin==1.2.0
sqlparse==0.4.1
whitenoise==6.2.0
